// Web-Back/FinalFernando/src/ModuloProductos/Domain/interfaces/IProductoData.ts
export default interface IProductoData  {
    idProducto: number;
    nombreProducto: string;
    descripcionProducto: string;
    precioProducto: number;
    stockProducto: number;
    imagenProducto: string;
    categoriaId: number;
    enPromocion: string; 
    categoriaNombre: string;
    marcaNombre: string;
    tallaNombre: string;
  }